+++ .gitignore
@@ -1,2 +1,3 @@
 /local.json
 .env
+/shared/local.json
