+++ web-api/src/invoice/invoice.service.ts
@@ -8,6 +8,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { Client } from '../common/entity/Clients';
 import { Log } from '../common/entity/Logs';
 import { Task } from './dto/invoice-request.dto';
+import { HelperInvoice } from './helper';
 
 @Injectable()
 export class InvoiceService {
@@ -17,26 +18,11 @@ export class InvoiceService {
     @InjectRepository(Log)
     private log: Repository<Log>,
     @InjectQueue('Invoice') private invoiceQueue: Queue,
+    private readonly helper: HelperInvoice,
   ) {}
 
-  getSumm(completedTasks: any[]) {
-    return completedTasks.reduce((accum: any, element: { cost: any }) => {
-      accum = element.cost + accum;
-      return accum;
-    }, 0);
-  }
-  getFormatedDate() {
-    const dateNow: Date = new Date();
-    return (
-      dateNow.getDate() +
-      '.' +
-      ('0' + (dateNow.getMonth() + 1)) +
-      '.' +
-      dateNow.getFullYear()
-    );
-  }
   async generate(
-    file,
+    file: File[],
     clientEmail: string,
     completedTasks: Task[],
     sendEmail: string,
@@ -50,11 +36,10 @@ export class InvoiceService {
       });
       throw new BadRequestException({ message: 'Invalid email' });
     }
-
     const { firstName, lastName } = client;
     const { name, address, scope } = client.company;
-    const summCost: number = this.getSumm(completedTasks);
-    const dateFormatDDMMYYYY = this.getFormatedDate();
+    const summCost: number = this.helper.getSummTask(completedTasks);
+    const dateFormatDDMMYYYY = this.helper.getFormatedDate();
     await this.log.save({
       email: clientEmail,
       firstName,
