+++ worker-email/config/default.json
@@ -6,7 +6,7 @@
     "passwordEmail": ""
   },
   "redisConfig": {
-    "host": "localhost",
+    "host": "cache_redis",
     "port": 6379,
     "password": "",
     "database": 2
