--- worker-email/src/interface/index.ts
@@ -1,12 +1,0 @@
-export class OwnerInfo {
-  email: string;
-  firstName: string;
-}
-
-export class ObjectForPDF {
-  sendEmail: string;
-  clientEmail: string;
-}
-export enum JobName {
-  GenerateAndSendInvoice = 'generateAndSendInvoice',
-}
