+++ worker-email/src/type/index.ts
@@ -1,0 +1,9 @@
+export type OwnerInfo = {
+  email: string;
+  firstName: string;
+};
+
+export type ObjectForPDF = {
+  sendEmail: string;
+  clientEmail: string;
+};
