+++ worker-pdf/config/default.json
@@ -1,6 +1,6 @@
 {
   "redisConfig": {
-    "host": "localhost",
+    "host": "cache_redis",
     "port": 6379,
     "password": "",
     "database": 2
