+++ worker-pdf/src/constants.ts
@@ -1,0 +1,3 @@
+export enum JobName {
+  GenerateAndSendInvoice = 'generateAndSendInvoice',
+}
