+++ worker-pdf/src/index.ts
@@ -1,7 +1,7 @@
 import { Process, Processor } from '@nestjs/bull';
 import { Logger } from '@nestjs/common';
 import { Job } from 'bull';
-import { JobName } from './interface';
+import { JobName } from './constants';
 import { generateInvoiceJob } from './jobs/generateInvoice';
 
 @Processor('Invoice')
