+++ worker-pdf/src/utils/generatePDF.ts
@@ -1,5 +1,5 @@
 import * as PDFDocument from 'pdfkit';
-import { ObjectForPDF, Owner } from '../interface';
+import { ObjectForPDF, Owner } from '../type';
 
 export async function generatePDF(
   objectForPDF: ObjectForPDF,
